Fixing uint vs. int in Heap.h

+++ mtl/Heap.h
@@ -175,7 +175,7 @@ class Heap {
                 indices[heap[i]] = std::numeric_limits<uint32_t>::max();
 
         heap.shrink(i - j);
-        for (int i = heap.size() / 2 - 1; i >= 0; i--)
+        for (int i = ((int)heap.size()) / 2 - 1; i >= 0; i--)
             percolateDown(i);
 
         assert(heapProperty());
